fixed setFoodLevel using an int parameter instead of a float

+++ TFC_Shared/src/TFC/Core/Player/FoodStatsTFC.java
@@ -250,7 +250,7 @@ public class FoodStatsTFC
 		return this.satisfaction;
 	}
 
-	public void setFoodLevel(int par1)
+	public void setFoodLevel(float par1)
 	{
 		this.stomachLevel = par1;
 	}
